:sparkles: feat: Utilsに関数を移動
